+++ packages/rocketh-signer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rocketh/signer
 
+## 0.11.22-testing.8
+
+### Patch Changes
+
+- Extra type generic
+
 ## 0.11.22-testing.7
 
 ### Patch Changes