+++ packages/rocketh-verifier/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/verifier
 
+## 0.11.24-testing.7
+
+### Patch Changes
+
+- Extra type generic
+- Updated dependencies
+  - rocketh@0.11.22-testing.7
+
 ## 0.11.24-testing.6
 
 ### Patch Changes